requirements.txt
# Core ML and Deep Learning
torch>=1.12.0
torchvision>=0.13.0
clip-by-openai>=1.0
transformers>=4.20.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.0.0
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.60.0
pyyaml>=6.0
click>=8.0.0

# Development (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900.0
